Fix height & whitespace issue in slider when hidden slides are longer
Hidden slides add to the height of slider. This reduces the height
of hidden slides, so that the slider doesn't have excess whitespace
below the visible slides.

The solution here is a bit hacky. We really shouldn't be applying CSS
on the aria-hidden attribute but have classes. Also, there should be
some way to compute the minimum height (to reduce flicker).

The overflow-y value prevents vertical scrolling of visible slides.